Refactor proxy options factory into its own class
