+++ src/FakeItEasy/Creation/IProxyOptionsFactory.cs
@@ -1,0 +1,9 @@
+namespace FakeItEasy.Creation
+{
+    using System;
+
+    internal interface IProxyOptionsFactory
+    {
+        IProxyOptions BuildProxyOptions(Type typeOfFake, Action<IFakeOptions>? optionsBuilder);
+    }
+}
